revert: remove plugin-a11y
Sadly, there is an upstream issue with the plugin and I'm not sure that
I'm able to get past it with global CSS (at least it appears not).

Therefore, I must turn it off in order for my builds to pass again.

I'll revist adding it back in once the issue is fixed upstream.